[Feat] placeholderData 이용 #118
